+++ src/aishell/cli.clj
@@ -26,7 +26,7 @@
             [aishell.util :as util]
             [aishell.update-check :as update-check]))
 
-(def version "3.19.0")
+(def version "3.20.0")
 
 (defn print-version []
   (println (str "aishell " version)))